+++ gpt_core/src/config.rs
@@ -1,10 +1,9 @@
+use crate::model::ModelVersion;
 use encrypt_config::PersistSource;
 use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use std::path::PathBuf;
 use url::Url;
-
-use crate::model::ModelVersion;
 
 /// The name of the application for config save.
 const NAME: &str = "chatGPT";
@@ -12,7 +11,7 @@ const NAME: &str = "chatGPT";
 /// The configuration for the OpenAI chatGPT API,
 /// which imlpements the `PersistSource` trait.
 /// `load` and `store` is provided.
-#[derive(Debug, Serialize, Deserialize, PersistSource)]
+#[derive(Debug, Serialize, Deserialize, PersistSource, Clone)]
 #[source(name = "gpt_cli/api_key.json")]
 pub struct Config {
     pub endpoint: Url,
@@ -20,16 +19,26 @@ pub struct Config {
     pub model: ModelVersion,
 }
 
+#[derive(Debug, Serialize, Deserialize, PersistSource, Clone)]
+#[source(name = "gpt_cli/api_keys.json")]
+pub struct Configs {
+    pub current: usize,
+    pub configs: Vec<Config>,
+}
+
 impl Display for Config {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(
             f,
-            "Use Config:\t{}\ndefault_model:\t{}\nEndpoint:\t{}\nAPI Key:\t{}",
-            Config::path().to_string_lossy(),
-            self.model,
-            self.endpoint,
-            self.api_key
+            "default_model: {}; Endpoint: {}; API Key: {}",
+            self.model, self.endpoint, self.api_key
         )
+    }
+}
+
+impl Display for Configs {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", &self.configs[self.current])
     }
 }
 
@@ -58,6 +67,40 @@ impl Config {
 
     pub fn store(&self) -> std::io::Result<()> {
         <Self as PersistSource>::store(self)
+    }
+}
+
+impl Default for Configs {
+    fn default() -> Self {
+        Self {
+            current: 0,
+            configs: vec![Config::default()],
+        }
+    }
+}
+
+impl Configs {
+    pub fn new() -> Self {
+        Self {
+            current: 0,
+            configs: vec![],
+        }
+    }
+
+    pub fn load() -> std::io::Result<Self> {
+        <Self as PersistSource>::load()
+    }
+
+    pub fn store(&self) -> std::io::Result<()> {
+        <Self as PersistSource>::store(self)
+    }
+
+    pub fn current(&self) -> &Config {
+        &self.configs[self.current]
+    }
+
+    pub fn current_mut(&mut self) -> &mut Config {
+        &mut self.configs[self.current]
     }
 }