+++ gpt_ui/src/cli/chat.rs
@@ -2,7 +2,7 @@ use super::{
     dialog::{confirm, input, select},
     error::Result,
 };
-use gpt_core::{chat::Chat, config::Config, msg::Role};
+use gpt_core::{chat::Chat, config::Configs, msg::Role};
 
 pub(crate) async fn chat() -> Result<()> {
     #[cfg(feature = "mock")]
@@ -25,9 +25,9 @@ pub(crate) async fn chat() -> Result<()> {
 
 pub(crate) async fn new_chat(mut chat: Chat) -> Result<()> {
     #[cfg(not(feature = "mock"))]
-    let config = Config::load().unwrap_or_default();
+    let config = Configs::load().unwrap_or_default().current().clone();
     #[cfg(feature = "mock")]
-    let config = Config {
+    let config = gpt_core::config::Config {
         endpoint: crate::MOCK_SERVER.parse().unwrap(),
         api_key: "".to_string(),
         model: gpt_core::model::ModelVersion::GPT4o,