+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - remove `dialoguer` and use `cliclack` for prompt
 - support DeepSeek (https://api.deepseek.com)
+- multiple config support
 
 ## [0.0.6] - 2024-10-19